In #13583, we removed the `equation_generation` logic, at it was unsound.
However, it means that `unify_gadt` can now fail because of non-injective type constructors.
This PR fixes that by first attempting to do normal unification in `unify_gadt` when there are type variables involved.