@garrigue 2024-10-31
I also added a new failure example, which is more interesting, as one can see that the types are not normally unifiable.