@garrigue 2024-11-06
I also tried removing the `closed_type_expr` check, but there is a tiny further slowdown.
(About 0.2%)