@Octachron 2024-11-06
Benchmarking on my GADT encoding that there are no [4x3 knight's tours](https://github.com/Octachron/typed-musings/blob/master/chess/tour.ml), there is a quite visible slowdown:

Benchmark 1: ./ocamlopt.opt -I stdlib/ -i tour.ml
 
|           | Time (mean) |  σ        | min | max |   
|-------- | --------------------|---------| ---------|--------| 
| 5.2.0   |   691.3 ms |  15.5 ms |    632.5 ms | 672.8 ms |
| trunk  | 751.6 ms  | 11.5 ms | 691.3 ms |  736.2 ms |
| This PR | 1.361 s | 0.013 s | 1.330 s |  1.396 s |