@gasche 2024-11-06
Yes, but (to help other people understand the discussion) as we discussed 2x is the worst case slowdown we can think of (for a code that spends all its time in GADT checking), so it's not actually so bad.