@garrigue 2024-11-07
I have added the suggested optimization.
Now the check is `penv.allow_recursive_equations || closed_type_expr ty2`.
I.e., if we are checking a counter-example, or if there are no free variables in the expected type, then no need to try normal unifcation.
I have added comments about the expected invariants.

To clarify the code, I plan to do the following refactoring in a separate PR:
* rename `allow_recursive_equations` to `force_refine` (or `counter_example` ?).
* in `Typecore`, remove `~refine` arguments, except in `unify_pat_types_return_equated_pairs`, which would become:
```ocaml
let unify_pat_types_return_equated_pairs ~refine loc penv ty ty' =
  try
    if refine || penv.force_refine then unify_gadt penv ty ty'
    else (unify !!penv ty ty'; nothing_equated)
  ...
```