@garrigue 2024-11-15
#13599 is an alternative to this one. I.e., it does not fix the essence of #13598, but I'm starting to think that it is as good as this to fix the side effects.