@Octachron 2024-11-20
I have sent a small commit that adds a label to `unify_gadt` to make the role of the two type expressions arguments more clearly asymmetrical in https://github.com/COCTI/ocaml/pull/16 .

My proposition for 5.3 would be to move forward with this PR as a safe fix, while possibly considering #13599 for later.