@garrigue 2024-11-28
Actually, if your read my comment above, the asymmetry is only in `unify_pat_types_return_equated_pairs`.
In particular all calls through `unify_pat_types_refine` are actually symmetrical (either normal unify, or both sides are reifiable).
Should I go forward and do the (small) refactoring suggested, as this makes all this clearer?