@Octachron 2024-11-28
Sorry I missed (or forgot) your comment about this point, sorry about this !

Removing the duplication between the pattern_env and the refine argument outside of sounds like a nice refactorization but I agree that we can leave this improvement for a subsequent PR.