@Octachron 2024-11-28
Cherry-picked on 5.3 as 32caed1e3, I have pulled along the refactoring from #13507 (in b73ea1e4) to reduce the distance between the trunk and 5.3 versions.